partial separation of validators and builders
